+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+### 0.11.5 - 2018-08-17
+
+- Improve non-object `state` handling
+- Improve `component.requestUpdate`
+
 ### 0.11.4 - 2018-08-17
 
-- Fix non-object `state` handling
+- Fix non-object state handling
 
 ### 0.11.3 - 2018-08-16